Created a skeleton Account class complete with the interfaces that will be needed on it.
It does not yet support a room since this will require the addition of more objects which I don't have time to do at the moment.

I've also added a "CreateTransaction" method to the DatabaseEngine - this will be useful later.

It seems that in .NET 2.0 the Command.CreateParameter method now requires a DbParameter to be passed to it rather than taking name value pairs. 

I've written two methods on the DatabaseEngine that handle creating parameter that are strings and long ints (int64).